implemented redux partially completed

+++ src/shared_components/LogoutButton.jsx
@@ -2,6 +2,9 @@
 import Button from '@mui/material/Button';
 import LogoutIcon from '@mui/icons-material/Logout';
 import { useNavigate } from "react-router-dom";
+import { settings } from '../config/config';
+import { clearActivities } from '../store/slices/activities';
+import { clearUserInfo } from '../store/slices/user';
 
 
 export default function LogoutButton() {
@@ -9,15 +12,17 @@ export default function LogoutButton() {
   const navigate = useNavigate();
 
   function userLogout() {
-    window.localStorage.clear();
-    navigate("/login");
+    if (settings.usingRedux) {
+      clearActivities();
+      clearUserInfo();
+    }
+    if (settings.usingLocalStorage) { window.localStorage.clear(); }
+    // navigate("/login");
   }
 
   return (
     <Button 
-      // variant="contained"
       startIcon={<LogoutIcon />}
-      // sx={{}}
       onClick={() => userLogout()}
       size="small"
     >
